% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_geom_ci.R
\name{xgx_geom_ci}
\alias{xgx_geom_ci}
\title{Plot data with mean and confidence intervals}
\usage{
xgx_geom_ci(percentile = 0.95, ...)
}
\arguments{
\item{percentile}{is the percentile for the confidence interval.  default is .95  should be between 0.5 and 1}

\item{...}{other arguments passed on to layer. These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
ggplot2 plot layer
}
\description{
Plot data with mean and confidence intervals
}
\examples{
library(ggplot2)  
data = data.frame(x = rep(c(1,2,3),each=20),
y = rep(c(1,2,3),each=20) + rnorm(60))
ggplot(data,aes(x=x,y=y)) + 
 xgx_geom_ci(percentile = .95)
}
