% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_labels_log10.R
\name{xgx_labels_log10}
\alias{xgx_labels_log10}
\title{Nice labels for log10.}
\usage{
xgx_labels_log10(breaks)
}
\arguments{
\item{breaks, }{breaks for the function}
}
\value{
either character or expression
}
\description{
Returns a set of labels for ggplot
}
\examples{
print(xgx_labels_log10(c(1e-5,1,1e5)))
}
