% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_scale_x_log10.R
\name{xgx_scale_x_log10}
\alias{xgx_scale_x_log10}
\title{log10 scales the x axis with a "pretty" set of breaks}
\usage{
xgx_scale_x_log10(breaks = xgx_breaks_log10,
  minor_breaks = xgx_log_breaks_minor, labels = xgx_labels_log10, ...)
}
\arguments{
\item{breaks}{major breaks, default is a function defined here}

\item{minor_breaks}{minor breaks, default is a function defined here}

\item{labels}{function for setting the labels, defined here}

\item{...}{other arguments passed to \code{ggplot2::scale_x_log10}}
}
\description{
\code{xgx_scale_x_log10} is similar to \code{ggplot2::scale_x_log10}.  
But it uses what we believe to be a nicer spacing and set of tick marks
it can be used the same as \code{scale_x_log10}
}
