% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_scale_y_reverselog10.R
\name{xgx_scale_y_reverselog10}
\alias{xgx_scale_y_reverselog10}
\title{Reverselog transform for the y scale.}
\usage{
xgx_scale_y_reverselog10(...)
}
\arguments{
\item{...}{other parameters passed to \code{ggplot2::scale_y_continuous}}
}
\description{
\code{xgx_scale_y_reverselog10} is designed to be used with data that approaches 100%.
A common example is receptor occupancy in drug development.  
It is used when you want even spacing between 90, 99, 99.9, etc.
}
\examples{
library(ggplot2)  
conc = 10^(seq(-3,3,by = 0.1))
EC50 = 1
data = data.frame(concentration  = conc, bound_receptor = 1*conc/(conc+EC50))
ggplot(data,aes(x=concentration,y=bound_receptor)) + 
geom_point() + 
 geom_line() + 
 xgx_scale_x_log10() +
 xgx_scale_y_reverselog10()
}
