% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_spaghetti.R
\name{xgx_spaghetti}
\alias{xgx_spaghetti}
\title{Make a spaghetti plot}
\usage{
xgx_spaghetti(data = NULL, mapping = aes(), ...,
  environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame, will be converted to one by fortify.}

\item{mapping}{As in ggplot2; Default list of aesthetic mappings to use for plot. Must define x, y, and group for xgx_spaghetti.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{environment}{If an variable defined in the aesthetic mapping is not found in the data, 
ggplot will look for it in this environment. It defaults to using the environment in which \code{\link{ggplot}} is called.}
}
\description{
\code{xgx_spaghetti} returns a spaghetti plot ggplot object
}
\examples{
library(ggplot2)  
time = rep(seq(1,10),5)
id = sort(rep(seq(1,5), 10))
conc = exp(-time)*sort(rep(rlnorm(5),10))

data = data.frame(time = time, concentration  = conc, id = id)
xgx_spaghetti(data = data, mapping = aes(x = time, y = concentration, group = id))
}
