% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_summarize_data.R
\name{xgx_summarize_data}
\alias{xgx_summarize_data}
\title{Check data for various issues}
\usage{
xgx_summarize_data(...)
}
\arguments{
\item{...}{parameters from \code{\link{xgx_check_data}}}
}
\description{
Calls \code{xgx_check_data}
}
